move capabilties e2e tests to members e2e tests
